Separate trace processing code from GUI code
Introduce a new package, trace/ptrace, that is responsible for
processing traces (creating spans, assigning tags etc) and split that
code out from the GUI code.

+++ cmd/gotraceui/trace_test.go
@@ -7,8 +7,8 @@ import (
 
 type nullProgresser struct{}
 
-func (nullProgresser) SetProgress(float32)      {}
-func (nullProgresser) SetProgressLossy(float32) {}
+func (nullProgresser) SetProgress(float64)      {}
+func (nullProgresser) SetProgressLossy(float64) {}
 
 func FuzzLoadTrace(f *testing.F) {
 	f.Fuzz(func(t *testing.T, in []byte) {
